+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1116](https://github.com/alleslabs/celatone-frontend/pull/1116) Add EVM contract details assets
 - [#1112](https://github.com/alleslabs/celatone-frontend/pull/1112) Add EVM contract details bytecodes
 - [#1109](https://github.com/alleslabs/celatone-frontend/pull/1109) Add EVM contract details overview
 - [#1106](https://github.com/alleslabs/celatone-frontend/pull/1106) Add EVM contract details HTML skeleton layout